third_party/blink/web_tests/fast/forms/select/select-max-length-expected.txt

CONSOLE WARNING: Unable to expand the option list to length 100001. The maximum allowed list length is 100000.
CONSOLE WARNING: Unable to expand the option list and set an option at index=100001. The maximum allowed list length is 100000.
CONSOLE WARNING: Unable to expand the option list and set an option at index=100001. The maximum allowed list length is 100000.
This test that setting HTMLSelectElement.length is capped to 100,000, and you can't add additional Option elements too.

PASS sel.length is 0
Trying: - sel.length = 100001;
PASS sel.length is 0
Trying: - sel.add(new Option, 0);
PASS sel.length is 1
Trying: - sel.length = 0;
PASS sel.length is 0
Index setter:
PASS sel[100001] = new Option(); sel.length is 0
PASS sel.options[100001] = new Option(); sel.length is 0
PASS successfullyParsed is true

TEST COMPLETE
